Fix build on recent FreeBSD (ie. upcoming FreeBSD 8.0; FreeBSD CURRENT).
Since March 9, 2009; the FreeBSD operating system base ships with libusb.  As a consequence, the devel/libusb ports that used to provide userland programs access to USB devices is useless and so marked as IGNORE for recent versions of FreeBSD.  This patch detects the FreeBSD version configure is run on and conditionally ignore the libusb detection if relevant.

Obtained from:	Ooo vcs.